feat(singleProducts): add product from single products page and redirect to cart IMPORTANT: addButton component removed
